# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News aggregator API documentation",
        "version": "1.0"
    },
    "paths": {
        "/api/v1/articles": {
            "get": {
                "tags": [
                    "articles"
                ],
                "summary": "articles resource",
                "operationId": "daffc9d151a0c7c956d1157d38e78859",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Global search keyword",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "published_at",
                        "in": "query",
                        "description": "The publish date of the article",
                        "required": false,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "The category of the article",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "The source of the article",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "number of articles per page",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page number",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "data": {
                                                "slug": "article1",
                                                "title": "article1",
                                                "description": "Laborum nesciunt ..",
                                                "content": "etur minus architect ..",
                                                "author": "r King",
                                                "category": "science",
                                                "source": "Newscred",
                                                "published_at": "2020-01-01"
                                            },
                                            "links": {
                                                "first": "articles?page=1",
                                                "next": "articles?page=2"
                                            },
                                            "meta": {
                                                "current_page": 1,
                                                "from": 1,
                                                "path": "/articles",
                                                "per_page": 10,
                                                "to": 10
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "unauthenticated."
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/articles/{article}": {
            "get": {
                "tags": [
                    "articles"
                ],
                "summary": "show article",
                "operationId": "fd791615b93c76e36ff869b8831d9568",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "article slug",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "slug": "totam",
                                            "title": "Totam",
                                            "description": "Vel quos occaecati",
                                            "content": "Iure et vero facere ",
                                            "author": "Prof. Deven Brei",
                                            "category": "Excepturi",
                                            "source": "Sint",
                                            "published_at": "1987-05-02"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "record not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "record not found."
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "user login",
                "operationId": "05d536c846963ff3d47fe782fa7f5435",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "swagger@test.com",
                                    "password": "password"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "data": {
                                                "user": {
                                                    "name": "john doe",
                                                    "email": "john@doe.com"
                                                },
                                                "token": "Rnw5mhtGF7e"
                                            },
                                            "message": "You have logged in successfully."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "failed login",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "The provided credentials don't match our records."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "user logout",
                "operationId": "90c6c29d208bd685486aca10d32d3dd4",
                "responses": {
                    "200": {
                        "description": "Logout success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "You have logged out successfully."
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}